requirements.txt

docx==0.2.4
docx2pdf==0.1.8
pdf2docx==0.5.9
pillow==12.1.1
python-docx==1.2.0
streamlit==1.54.0
